+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/action/ShowRichPresenceAction.kt
@@ -2,12 +2,9 @@ package io.github.pandier.intellijdiscordrp.action
 
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.application.EDT
 import com.intellij.openapi.project.DumbAwareToggleAction
 import io.github.pandier.intellijdiscordrp.service.DiscordService
 import io.github.pandier.intellijdiscordrp.settings.project.discordProjectSettingsComponent
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
 
 class ShowRichPresenceAction : DumbAwareToggleAction() {
 
@@ -19,15 +16,14 @@ class ShowRichPresenceAction : DumbAwareToggleAction() {
         e.presentation.isEnabledAndVisible = e.project != null
     }
 
-    override fun isSelected(e: AnActionEvent): Boolean =
-        e.project?.discordProjectSettingsComponent?.state?.showRichPresence ?: true
+    override fun isSelected(e: AnActionEvent): Boolean {
+        return e.project?.discordProjectSettingsComponent?.state?.showRichPresence ?: true
+    }
 
     override fun setSelected(e: AnActionEvent, state: Boolean) {
         e.project?.discordProjectSettingsComponent?.state?.showRichPresence = state
 
         val discordService = DiscordService.getInstance()
-        discordService.scope.launch(Dispatchers.EDT) {
-            discordService.updateActivity()
-        }
+        discordService.updateBackground()
     }
 }