+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/listener/activity/RichPresenceProjectListenerActivity.kt
@@ -1,0 +1,13 @@
+package io.github.pandier.intellijdiscordrp.listener.activity
+
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.startup.ProjectActivity
+import io.github.pandier.intellijdiscordrp.service.DiscordService
+
+class RichPresenceProjectListenerActivity : ProjectActivity {
+
+    override suspend fun execute(project: Project) {
+        val discordService = DiscordService.getInstance()
+        discordService.switchProject(project)
+    }
+}