+++ CHANGELOG.md
@@ -6,19 +6,76 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2024-05-09
+
 ### Added
 
+- File types for Svelte, Nuxt config (`nuxt.config.ts`), NPM package (`package.json`, `package-lock.json`), TypeScript config (`tsconfig.json`)
+
+### Changed
+
+- Move Rich Presence configuration options to root
+
+## [1.2.1] - 2024-05-08
+
+### Added
+
+- Add `mjs` extension support for JavaScript
+
+### Fixed
+
+- `ConnectionException` on Unix machines
+
+## [1.2.0] - 2024-05-02
+
+### Added
+
+- "Show elapsed time" checkbox that toggles the timestamp for a display mode
+- Text field length validation
+- File type for diff preview
+
+### Changed
+
+- Improved error handling
+- KPresence is now used for back-end
+
+### Fixed
+
+- Emojis are now working in Rich Presence
+
+## [1.1.1] - 2024-04-20
+
+### Fixed
+
+- Remove usage of experimental api
+
+## [1.1.0] - 2024-04-20
+
+### Added
+
+- `{file_dir_name}` variable that displays the directory name of the edited file
+- `{app_version}` variable that displays the version of the IDE
+- Progress indicator and notification for reconnecting action
+- New icons for Dart, images
+- Improved settings user interface
+- Support for remaining JetBrains IDEs
 - Setting for changing display mode per project (Tools -> Discord Rich Presence -> Change Display Mode in Project)
 
   - The default value can be configured in global settings
-
 - Added application display mode
 
 ### Changed
 
+- `{file_path}` now uses path relative to content root
+- `{app_full_name}` now uses edition instead of version
+- Apply settings reconnecting only when custom application id is modified
 - Minimum supported IntelliJ version is now 2023.1
 - Display mode settings are now grouped into tabs
 - All actions are now dumb aware
+
+### Fixed
+
+- Blank fields causing the activity to not show up
 
 ## [1.0.2] - 2024-03-29
 
@@ -36,7 +93,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release
 
-[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.0.2...HEAD
+[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.2.1...v1.3.0
+[1.2.1]: https://github.com/pandier/intellij-discord-rp/compare/v1.2.0...v1.2.1
+[1.2.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.1.1...v1.2.0
+[1.1.1]: https://github.com/pandier/intellij-discord-rp/compare/v1.1.0...v1.1.1
+[1.1.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.0.2...v1.1.0
 [1.0.2]: https://github.com/pandier/intellij-discord-rp/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/pandier/intellij-discord-rp/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/pandier/intellij-discord-rp/commits/v1.0.0